+++ src/types/form-element.ts
@@ -348,6 +348,13 @@ export interface DateTimeOptions {
    * @type {string}
    */
   value?: string;
+
+  /**
+   * Is Transfrom to UTC
+   *
+   * @type {boolean}
+   */
+  isUseLocalTime: boolean;
 }
 
 /**